refactor: use BigDecimal for volatility multiplier, add test rationale comments, update README diagrams
- Change StockPortIn.setVolatilityMultiplier(double) to BigDecimal for
  domain model consistency with other numeric fields
- Change StockService volatilityMultiplier field from
  AtomicReference<Double> to AtomicReference<BigDecimal>
- Add class-level test rationale comments to StockServiceTest and
  StockPriceTickSchedulerTest per project testing standards
- Update README Chaos Level Management diagram to show StockPortIn as
  explicit cross-module participant for setVolatilityMultiplier
- Add README prose explaining volatility multiplier push from
  ChaosService to StockService

+++ stonks_java/src/main/java/dev/pollito/stonks_java/chaos/application/service/ChaosService.java
@@ -8,6 +8,7 @@ import dev.pollito.stonks_java.chaos.domain.ChaosEventTriggered;
 import dev.pollito.stonks_java.chaos.domain.ChaosLevel;
 import dev.pollito.stonks_java.news.application.port.in.NewsPortIn;
 import dev.pollito.stonks_java.stock.application.port.in.StockPortIn;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 import lombok.RequiredArgsConstructor;
@@ -53,6 +54,6 @@ public class ChaosService implements ChaosPortIn {
   @Override
   public void setLevel(ChaosLevel level) {
     currentLevel.set(level);
-    stockPortIn.setVolatilityMultiplier(level.getVolatilityMultiplier());
+    stockPortIn.setVolatilityMultiplier(BigDecimal.valueOf(level.getVolatilityMultiplier()));
   }
 }
